bashrc: fine-tuned install.sh and added ubuntu_server specific .bashrc

+++ install.sh
@@ -10,17 +10,28 @@
 
 # NVIM
 sudo apt install curl
-mkdir -p nvim/dein/
-cp -r ./nvim/ ~/.config/nvim/
-sudo apt install neovim
-curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > ./nvim/dein/installer.sh
-sh ./nvim/dein/installer.sh ~/.config/nvim/dein/
+echo "\nInstall NVIM..."
+if ! [ -d ~/.config/nvim/dein/ ]; then
+	mkdir -p ~/.config/nvim/dein/
+	curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > ./nvim/dein/installer.sh
+	sh ./nvim/dein/installer.sh ~/.config/nvim/dein/
+fi
+
+if ! [ -f /usr/bin/nvim ]; then
+	sudo apt install neovim
+	cp -r ./nvim/ ~/.config/nvim/
+fi
+
 
 # VIM
-mkdir ~/.vim/
-cp -r ./nvim/* ~/.vim/
-rm ~/.vim/init.vim
-cp ./nvim/init.vim ~/.vimrc
+echo "\nInstall VIM..."
+if ! [ -d ~/.vim ]; then
+	mkdir ~/.vim/
+	cp -r ./nvim/* ~/.vim/
+	rm ~/.vim/init.vim
+fi
+cp ./nvim/init_vimrc.vim ~/.vimrc
 
 # BASHRC
-cp ./bashrc_server_settings/.bashrc ~/.bashrc
+echo "\nInstall .bashrc config..."
+cp ./bashrc/.bashrc_server ~/.bashrc
